refactor(github): split teamwork tab into separate components
- Move SectionTitle, TeamworkTable, and TeamworkInteractionsTable into individual files
- Create new index file to export teamwork components
- Update TeamworkTab to use new components
- Extract colors and styles into TeamworkTheme
