+++ src/components/github/analysis/components/teamwork/TeamworkTheme.tsx
@@ -1,0 +1,14 @@
+/**
+ * Colors and theme settings used across teamwork components
+ */
+export const TeamworkColors = {
+	main: "#EC4899", // pink
+	secondary: "#8B5CF6", // purple
+	tertiary: "#F59E0B", // amber
+	light: "rgba(236, 72, 153, 0.1)",
+	lighter: "rgba(236, 72, 153, 0.05)",
+	gradient: "linear-gradient(90deg, #EC4899 0%, #D946EF 100%)",
+	gradientSecondary: "linear-gradient(90deg, #8B5CF6 0%, #A78BFA 100%)",
+	primaryText: "#BE185D",
+	secondaryText: "#5B21B6",
+};
