+++ src/components/github/analysis/components/teamwork/index.ts
@@ -1,0 +1,4 @@
+export { default as SectionTitle } from './SectionTitle';
+export { default as TeamworkTable } from './TeamworkTable';
+export { default as TeamworkInteractionsTable } from './TeamworkInteractionsTable';
+export { TeamworkColors } from './TeamworkTheme';
